refactor: centralize modal clone token resolution
