+++ packages/modal-infra/tests/test_image_builder_v2.py
@@ -383,7 +383,7 @@ class TestBuildRepoImage:
 
         with (
             patch("src.scheduler.image_builder.validate_control_plane_url", return_value=True),
-            patch("src.scheduler.image_builder._generate_clone_token", return_value="gh-token"),
+            patch("src.scheduler.image_builder.resolve_clone_token", return_value="gh-token"),
             patch("src.sandbox.manager.SandboxManager", return_value=manager),
             patch(
                 "src.scheduler.image_builder._callback_with_retry",
@@ -415,7 +415,7 @@ class TestBuildRepoImage:
 
         with (
             patch("src.scheduler.image_builder.validate_control_plane_url", return_value=True),
-            patch("src.scheduler.image_builder._generate_clone_token", return_value="gh-token"),
+            patch("src.scheduler.image_builder.resolve_clone_token", return_value="gh-token"),
             patch("src.sandbox.manager.SandboxManager", return_value=manager),
             patch(
                 "src.scheduler.image_builder._callback_with_retry",
@@ -444,7 +444,7 @@ class TestBuildRepoImage:
 
         with (
             patch("src.scheduler.image_builder.validate_control_plane_url", return_value=True),
-            patch("src.scheduler.image_builder._generate_clone_token", return_value="gh-token"),
+            patch("src.scheduler.image_builder.resolve_clone_token", return_value="gh-token"),
             patch("src.sandbox.manager.SandboxManager", return_value=manager),
             patch(
                 "src.scheduler.image_builder._callback_with_retry",
@@ -474,7 +474,7 @@ class TestBuildRepoImage:
 
         with (
             patch("src.scheduler.image_builder.validate_control_plane_url", return_value=True),
-            patch("src.scheduler.image_builder._generate_clone_token", return_value="gh-token"),
+            patch("src.scheduler.image_builder.resolve_clone_token", return_value="gh-token"),
             patch("src.sandbox.manager.SandboxManager", return_value=manager),
             patch(
                 "src.scheduler.image_builder._callback_with_retry",
@@ -524,7 +524,7 @@ class TestBuildRepoImage:
 
         with (
             patch("src.scheduler.image_builder.validate_control_plane_url", return_value=True),
-            patch("src.scheduler.image_builder._generate_clone_token", return_value="gh-token"),
+            patch("src.scheduler.image_builder.resolve_clone_token", return_value="gh-token"),
             patch("src.sandbox.manager.SandboxManager", return_value=manager),
             patch(
                 "src.scheduler.image_builder._callback_with_retry",
